+++ Changes.md
@@ -3,6 +3,7 @@
   - Updated framework references to 4.6.1 from 4.5
   - Add SetExtendedProperty method to LogContext
   - Fix a couple of brittle tests
+  - Sign the assemblies - except LogBridge.UmbracoLog4Net
 
 ## Version 1.3